// smsf-masterdata-ui/smsf-masterdata-ui/src/app/masterdata-services/model/clientServiceLevel.ts
/**
 * Automation API provider
 * Automation provides template of this description, and anyone may modify it in your own project.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ClientServiceLevel { 
    code?: string;
    ety?: number;
    id?: string;
    name?: string;
    parentId?: string;
    processId?: string;
}